+++ CHANGELOG.md
@@ -1097,6 +1097,7 @@ activity and on its child transactions (which can be actuals, refunds, and adjus
 - Refactor the importers (including associated uploads controllers and views) to use a unified approach for all entities being imported
 - Ensure `Forecast.set_value` always returns a forecast or nil; this will ensure that uploaded forecasts are correctly displayed back to the user on the success page
 - Add Level B budget bulk upload functionality - form with errors/confirmation view; link in top nav
+- Include refunds and adjustments in the calculation of an activity's total spend (previously only actuals were included)
 
 [unreleased]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-117...HEAD
 [release-117]: https://github.com/UKGovernmentBEIS/beis-report-official-development-assistance/compare/release-116...release-117